+++ crates/migrator/src/migrations.rs
@@ -1,0 +1,5 @@
+pub(crate) mod m_0001 {
+    mod settings;
+
+    pub(crate) use settings::rename_status_bar_show;
+}