+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Open file by absolute path in file finder (#212)
 - Add an option to hide the unsaved indicator (#266) by @bux
 - Make LSP auto-updating an opt-in setting ((#267) by @bux
+- Add option to move status bar to top of window
 
 [poll]: https://github.com/lilith/zed/commit/0334469a57a20586b28b86187028acd36559a9d3
 [kintsugi]: https://github.com/giacomocavalieri/kintsugi-zed
@@ -92,6 +93,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use SSH nicknames in display names (zed#53103)
 - Bump tree-sitter for fix to wasm loading of grammars w/ reserved words (zed#52856)
 - Update .gitattributes (#226) by @ivankopylov6603
+- Make status_bar.show a non-experimental setting
 
 ## [1.2.1] - 2026-03-26